Add 'stretch' program; make Visual Studio files usable.
The 'stretch' program allows for stretching/shrinking of spaces between
bars (measures) in tracks (without affecting tempo), as well as for
stretching/shrinking the duration (tempo) of the tracks themselves.

Also, added a Visual Studio solution containing the project file for the
midifile library as well as several projects for client programs
(including 'stretch').  More project files for programs can be added
later.